move FileSystem implementation from header to source
