+++ gframe/config.h
@@ -62,22 +62,6 @@ inline T myclamp(T v, T lo, T hi) {
 	return (v < lo) ? lo : (hi < v) ? hi : v;
 }
 
-inline FILE* mywfopen(const wchar_t* filename, const char* mode) {
-	FILE* fp{};
-#ifdef _WIN32
-	wchar_t wmode[20]{};
-	BufferIO::CopyCharArray(mode, wmode);
-	fp = _wfopen(filename, wmode);
-#else
-	char fname[1024]{};
-	BufferIO::EncodeUTF8(filename, fname);
-	fp = std::fopen(fname, mode);
-#endif
-	return fp;
-}
-
-#define myfopen std::fopen
-
 #include <irrlicht.h>
 
 constexpr uint16_t PRO_VERSION = 0x1362;
