+++ gframe/data_manager.cpp
@@ -1,6 +1,7 @@
 #include "data_manager.h"
 #include "game.h"
 #include "client_card.h"
+#include "file_system.h"
 
 namespace ygo {
