+++ gframe/game.cpp
@@ -1,6 +1,6 @@
 #include "config.h"
 #include "game.h"
-#include "myfilesystem.h"
+#include "file_system.h"
 #include "image_manager.h"
 #include "data_manager.h"
 #include "deck_manager.h"
